Make parse a module, add attrs submodule
